note on sdn-node; minion->node
After starting openshift-sdn-node I checked status and saw errors, and saw that openshift-node wasn't started. This was alarming. I expect users to do the same, so reassure them that this is normal.

Also, "Minion" has been deprecated in favor of "Node".
And I thought "public" is a confusing word for a NIC IP in the context of EC2/OpenStack.