docs/01-prerequisites: No pre-formatting for "Configure everything ..."
We added this line for context, it's not part of the 'aws configure
...' session.

Also use console syntax highlighting here.  GitHub uses Linguist for
syntax highlighting [1], and Linguist defines two related grammars:

* Shell (alias 'sh') [2], which applies to POSIX and similar shell
  languages.

* ShellSession (alias 'console') [3], which extends the Shell grammar
  with support for prompts and command output as you would usually see
  in an interactive shell session.

The 'aws configure' invocation includes output, so it deserves
'console' highlighting.  The 'export' command could go either way, but
it had a prompt before this commit, so I've gone with console
highlighting there too (and it's nice to have a consistent approach).

[1]: https://help.github.com/articles/creating-and-highlighting-code-blocks/#syntax-highlighting
[2]: https://github.com/github/linguist/blob/v7.1.3/lib/linguist/languages.yml#L4465-L4526
[3]: https://github.com/github/linguist/blob/v7.1.3/lib/linguist/languages.yml#L4527-L4538