docs/01-prerequisites: Use ~/.ssh/*.pub for pubkeys
These don't have to be RSA keys.

Also point to the machine-config operator docs for SSH key management,
to relieve any concerns that the previous note paragraph may have
raised ;).  It's still a reasonable idea to provide your SSH key at
install time so you can access the bootstrap machine for debugging,
but as far as production machines go, it won't matter either way (and
you'll only get SSH keys on those nodes if they come up alive enough
to get a functional machine-config daemon).

Another wrinkle with the current installer is that if it only finds
one ~/.ssh/*.pub file, it will just use that without prompting the
user to choose between that public key and <none>.  This also seems
too peripheral to be worth calling out in the docs.